10주차 코드 PR

+++ src/main/java/poomasi/domain/farm/dto/FarmResponse.java
@@ -3,15 +3,15 @@ package poomasi.domain.farm.dto;
 import poomasi.domain.farm.entity.Farm;
 
 
-public record FarmResponse( // FIXME: 사용자 정보 추가 및 설명/전화번호 추가
-                            Long id,
-                            String name,
-                            String address,
-                            String addressDetail,
-                            Double latitude,
-                            Double longitude,
-                            String description,
-                            Long experiencePrice
+public record FarmResponse(
+        Long id,
+        String name,
+        String address,
+        String addressDetail,
+        Double latitude,
+        Double longitude,
+        String description,
+        int experiencePrice
 ) {
     public static FarmResponse fromEntity(Farm farm) {
         return new FarmResponse(
